requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# Data Processing
pandas
pymupdf
opencv-python-headless
Pillow

# Text Splitting
langchain-text-splitters

# Vector Database (Pinecone - cloud)
pinecone

# Embeddings
sentence-transformers

# Multimodal (CLIP)
transformers
torch
torchvision

# HTTP client (for Ollama)
requests

# Neo4j graph database driver
neo4j
